requirements.txt
dagster==1.3.13
dagit==1.3.13
pandas==2.0.1
pyarrow==11.0.0
pydantic!=1.10.7,<2.0.0
sqlalchemy==2.0.12
plotly==5.14.1
fastapi==0.95.1
uvicorn==0.22.0
click==8.1.3
jupyter==1.0.0
nbconvert==7.3.1
pytest==7.3.1
pyyaml==6.0
requests==2.31.0
pendulum==2.1.2
numpy==1.24.3
dagster-graphql==1.3.13